reorganiza modularização de helper functions
